requirements.txt
numpy>=1.23.4
pandas>=1.5.0
openai>=0.25.0
python-dotenv>=0.21.0
transformers>=4.25.1
pypdf2>=3.0.1